flask/requirements.txt
Flask==2.1.0
Werkzeug==2.0.2
mysql-connector-python==8.0.26
PyYAML==5.4.1
